backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
git+https://github.com/openai/whisper.git
python-dotenv==1.0.0
celery[redis]==5.3.6
redis==5.0.1
boto3==1.29.3
pydantic>=2.0.0
elevenlabs==0.2.26
google-generativeai>=0.3.0
moviepy==1.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
typing-extensions>=4.8.0
torch
numpy
tqdm 